app/requirements.txt
alembic==1.13.3
black==24.8.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
defusedxml==0.7.1
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
Mako==1.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
numpy==2.1.1
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.35
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
gunicorn==21.2.0
signxml==3.2.2
flask-restx